actualizo el codigo

+++ Fase 2/eclipse-workspace/POS_Sistema/src/pos/model/SaleItem.java
@@ -26,8 +26,7 @@ public class SaleItem {
         this.quantity = quantity;
     }
 
-    // --- Calcular total por item ---
-    public double getTotal() {
+    public int getTotal() {
         return product.getPrice() * quantity;
     }
